requirements.txt
pyttsx3==2.99
SpeechRecognition==3.14.3
PyAudio==0.2.14
google-genai==1.38.0
python-dotenv==1.1.1